+++ mobile/android/android-components/components/concept/engine/src/main/java/mozilla/components/concept/engine/webextension/WebExtension.kt
@@ -306,6 +306,11 @@ abstract class Port(val engineSession: EngineSession? = null) {
     abstract fun name(): String
 
     /**
+     * Returns the URL of the port sender.
+     */
+    abstract fun senderUrl(): String
+
+    /**
      * Disconnects this port.
      */
     abstract fun disconnect()